% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adj2list}
\alias{adj2list}
\title{Convert adjacency matrix to adjacency list}
\usage{
adj2list(X)
}
\arguments{
\item{X}{adjacency matrix}
}
\description{
Convert adjacency matrix to adjacency list
}

