% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDSEstimators.R
\name{HH_mean}
\alias{HH_mean}
\title{HH estimator}
\usage{
HH_mean(chain, trait.vec, inclusion.probability)
}
\arguments{
\item{chain}{a sequence of node indices that represent the referral process.}

\item{trait.vec}{is a vector of size number of nodes in the network with corresponding trait values.}

\item{inclusion.probability}{is a vector of size number of nodes in the network with corresponding stationary probabilities.}
}
\description{
HH estimator
}

