% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{rdssim}
\alias{rdssim}
\title{Simulates referral chains/trees for Respondent-driven sampling}
\usage{
rdssim(adjlist, rType, wRreplacement, nSamples, nReferrals, seedNode, rseed)
}
\arguments{
\item{adjlist}{adjacency list of the input social network. It assumes that adjlist represents an undirected network. Refer to \code{adj2list}.}

\item{rType}{referral type.}

\item{wRreplacement}{if TRUE, then the referral chain/tree will be done with replacement.}

\item{nSamples}{number of samples to be collected.}

\item{nReferrals}{number of referral per vertex.}

\item{seedNode}{the starting vertex.}

\item{rseed}{random seed for random number generators.}
}
\description{
Simulates referral chains/trees for Respondent-driven sampling
}

