% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdssimMarkov}
\alias{rdssimMarkov}
\title{Respondent-driven sampling simulator}
\usage{
rdssimMarkov(rcpp_adjlist, rType, nSamples, nReferrals = 1L, seedNode = 1L,
  rseed = 1L)
}
\arguments{
\item{rcpp_adjlist}{adjacancy list. Refert to \code{\link{adj2list}}}

\item{rType}{referral type}

\item{nSamples}{numeber of sample to be collected}

\item{nReferrals}{number of referrals per person}

\item{seedNode}{starting seed node (person)}

\item{rseed}{random seed for the random number generator that generates the random referrals}
}
\description{
Respondent-driven sampling simulator
}

