% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDSEstimators.R
\name{sample_mean}
\alias{sample_mean}
\title{Sample mean estimator}
\usage{
sample_mean(chain, trait.vec)
}
\arguments{
\item{chain}{a sequence of node indices that represent the referral process.}

\item{trait.vec}{is a vector of size number of nodes in the network with corresponding trait values.}
}
\description{
Sample mean estimator
}

