% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{adj2list}
\alias{adj2list}
\title{Converts an adjacency matrix into the adjacency list}
\usage{
adj2list(A)
}
\arguments{
\item{A}{a symmetric unweighted adjacency matrix.}
}
\description{
Converts an adjacency matrix into the adjacency list
}

