% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdssimMarkov}
\alias{rdssimMarkov}
\title{Simulates referral chains/trees for Respondend-driven sampling}
\usage{
rdssimMarkov(adjlist, rType, nSamples, nReferrals = 1L, seedNode = 1L,
  rseed = 1L)
}
\arguments{
\item{adjlist}{adjacency list of the input social network. It assumes that adjlist represents an undirected network. Refer to \code{adj2list}.}

\item{rType}{referral type.}

\item{nSamples}{number of samples to be collected.}

\item{nReferrals}{number of referral per node (person).}

\item{seedNode}{the starting node (person).}

\item{rseed}{random seed for random number generators.}
}
\description{
Simulates referral chains/trees for Respondend-driven sampling
}

